fixed it

+++ code/modules/lighting/lighting_corner.dm
@@ -31,6 +31,7 @@
 	var/below_r = 0
 	var/below_g = 0
 	var/below_b = 0
+	var/is_lit_from_below = FALSE
 
 	//additive light values
 	var/add_r = 0
@@ -90,7 +91,7 @@
 			master.lighting_corner_SE = src
 
 /datum/lighting_corner/proc/self_destruct_if_idle()
-	if (!LAZYLEN(affecting))
+	if (!LAZYLEN(affecting) && !is_lit_from_below)
 		qdel(src, force = TRUE)
 
 /datum/lighting_corner/proc/vis_update()
@@ -101,8 +102,8 @@
 	for (var/datum/light_source/light_source as anything in affecting)
 		light_source.recalc_corner(src)
 
-#define CHECK_BELOW_LUM(Tt, corner) \
-	if((T = Tt?.below)) { \
+#define UPDATE_ABOVE_LUM(Tt, corner) \
+	if((T = Tt?.above)) { \
 		if(T.lighting_object) { \
 			C = T.##corner; \
 			if(!C) { \
@@ -136,10 +137,10 @@
 	#ifdef ZMIMIC_LIGHT_BLEED
 	var/turf/T
 	var/datum/lighting_corner/C
-	CHECK_BELOW_LUM(master_NE, lighting_corner_NE)
-	CHECK_BELOW_LUM(master_SE, lighting_corner_SE)
-	CHECK_BELOW_LUM(master_SW, lighting_corner_SW)
-	CHECK_BELOW_LUM(master_NW, lighting_corner_NW)
+	UPDATE_ABOVE_LUM(master_NE, lighting_corner_NE)
+	UPDATE_ABOVE_LUM(master_SE, lighting_corner_SE)
+	UPDATE_ABOVE_LUM(master_SW, lighting_corner_SW)
+	UPDATE_ABOVE_LUM(master_NW, lighting_corner_NW)
 	#endif
 
 	if (!needs_update)
@@ -153,6 +154,7 @@
 	below_r += delta_r
 	below_g += delta_g
 	below_b += delta_b
+	is_lit_from_below = below_r || below_g || below_b
 
 	UPDATE_SUM_LUM(r)
 	UPDATE_SUM_LUM(g)
@@ -242,5 +244,5 @@
 
 	return ..()
 
-#undef CHECK_BELOW_LUM
+#undef UPDATE_ABOVE_LUM
 #undef UPDATE_SUM_LUM
